@@ -7,21 +7,20 @@ git_source(:github) do |repo_name|
   "https://github.com/#{repo_name}.git"
 end
 
-gem 'puma', '~> 5'
-gem 'rails', '~> 5.2'
-# seg fault with sassc 2.3.0
+gem 'puma', '~> 6'
+gem 'rails', '~> 6.1'
 gem 'sassc', '~> 2.4'
 gem 'sassc-rails'
 gem 'uglifier'
 
-gem 'friendly_id', '~> 5.4.0'
+gem 'friendly_id'
 
-gem 'coffee-rails'
 gem 'jbuilder'
 # Use Redis adapter to run Action Cable in production
-# gem 'redis', '~> 3.0'
+gem 'redis', '~> 4'
 # Use ActiveModel has_secure_password
 gem 'bcrypt'
+gem 'deep_cloneable', '~> 3'
 
 # custom
 gem 'cancancan'
@@ -29,76 +28,86 @@ gem 'devise'
 gem 'exception_notification'
 gem 'foundation-icons-sass-rails'
 gem 'foundation-rails', '~> 6'
-gem 'haml-rails', '~> 2.0'
-gem 'i18n-js'
+gem 'geocoder'
+gem 'haml-rails'
+gem 'i18n-js', '~> 3' # breaking changes with v4
 gem 'jquery-rails'
 gem 'jquery-turbolinks'
-gem 'jquery-ui-rails'
+gem 'jquery-ui-rails', '~> 7', git: 'https://github.com/jquery-ui-rails/jquery-ui-rails.git', branch: 'master', ref: '27a942cfa3686716ff85a3cb60f1f885a00e35dc'
 gem 'kaminari'
 gem 'mysql2'
 gem 'rack-attack'
 gem 'rack-cors'
 gem 'secure_headers'
 gem 'select2-rails'
-gem 'settingslogic'
 gem 'simple_form'
-gem 'sprockets', '>= 3.0.0'
+gem 'sprockets', '~> 4'
 gem 'sprockets-es6'
 
-gem 'image_processing', '~> 1.2'
+gem 'image_processing'
 gem 'mini_racer', platforms: :ruby
+gem 'nokogiri'
 
+gem 'acts-as-taggable-on', '~> 9'
 gem 'color-generator'
-
-gem 'tinymce-rails'
-
-gem 'acts-as-taggable-on', '~> 7'
-
+gem 'mini_exiftool'
+gem 'mini_magick'
+gem 'mobility', '~> 1.2'
+gem 'rubyzip'
+gem 'tinymce-rails', '~> 5'
 gem 'turnout'
 
-gem 'mobility', '~> 1.1.2'
+gem 'net-smtp'
 
 group :staging, :production do
-  # gem 'pg'
   gem 'passenger'
 end
 
 gem 'coveralls_reborn', '~> 0', require: false
 
-group :development, :test, :localtest do
+group :development, :test do
+  gem 'bundler-audit', require: false
+  gem 'rubocop'
+  gem 'rubocop-rspec'
+  gem 'ruby_audit', require: false
+
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: %i[mri mingw x64_mingw]
-  # Adds support for Capybara system testing and selenium driver
   gem 'capybara', '~> 3'
   gem 'capybara-chromedriver-logger'
   gem 'database_cleaner'
   gem 'factory_bot_rails'
   gem 'json-schema'
   gem 'rails-controller-testing'
+  gem 'rspec'
   gem 'rspec_junit_formatter'
   gem 'rspec-rails'
+  gem 'selenium-webdriver'
+  gem 'shoulda-matchers'
   gem 'simplecov'
-  gem 'webdrivers', '~> 4.0'
+  gem 'simplecov-small-badge', require: false
+  gem 'webdrivers', '~> 5'
+
   gem 'webmock'
 end
 
 group :development do
   # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
   gem 'listen', '~> 3'
-  gem 'web-console', '~> 3'
+  gem 'web-console', '~> 4'
 
-  gem 'capistrano', '~> 3.15', require: false
+  gem 'capistrano', '~> 3', require: false
   gem 'capistrano-bundler'
   gem 'capistrano-passenger'
-  gem 'capistrano-rails', '~> 1.3', require: false
+  gem 'capistrano-rails', require: false
   gem 'capistrano-rvm'
 
   gem 'letter_opener'
 
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
-  gem 'spring-watcher-listen', '~> 2.0.0'
 
   gem 'brakeman', require: false
-  gem 'rubycritic', require: false
 end
+
+gem 'activerecord-session_store', '~> 2.1'
