@@ -5,7 +5,6 @@ class MapsController < ApplicationController
 
   before_action :redirect_to_friendly_id, only: %i[show]
 
-
   # GET /maps
   # GET /maps.json
   def index
@@ -16,14 +15,35 @@ class MapsController < ApplicationController
   # GET /maps/1.json
   def show
     @maps = Map.sorted.by_user(current_user)
+
     if @map
       @map_layers = @map.layers
+
+      @places = @map_layers.flat_map(&:places)
+      @places_with_dates = @places.reject { |place| place.startdate.nil? && place.enddate.nil? }
+
+      # timeline calculation, for now on a yearly basis
+      @minyear = @places.reject { |place| place.startdate.nil? }.min_by { |place| place.startdate.year }&.startdate&.year || Date.today.year
+      @maxyear = @places.reject { |place| place.enddate.nil? }.max_by { |place| place.enddate.year }&.enddate&.year || Date.today.year
+
+      # make a hash, where the key is a single year and it contains all places that are active in that year
+      @places_by_year = {}
+      @places_with_dates.each do |place|
+        startyear = place.startdate.nil? ? @minyear : place.startdate.year
+        endyear = place.enddate.nil? ? startyear : place.enddate.year
+        (startyear..endyear).each do |year|
+          @places_by_year[year.to_i] ||= []
+          @places_by_year[year.to_i] << place
+        end
+      end
+      @timespan = @maxyear - @minyear
+
       respond_to do |format|
         format.html { render :show }
         format.json { render :show, filename: "orte-map-#{@map.title.parameterize}-#{I18n.l Date.today}.json" }
       end
     else
-      redirect_to maps_path, notice: "Sorry, this map could not be found."
+      redirect_to maps_path, notice: 'Sorry, this map could not be found.'
     end
   end
 
@@ -82,23 +102,19 @@ class MapsController < ApplicationController
   private
 
   def redirect_to_friendly_id
-
     # If an old id or a numeric id was used to find the record, then
     # the request path will not match the post_path, and we should do
     # a 301 redirect that uses the current friendly id.
-    if @map && request.path != map_path(@map) && request.format == 'html'
-      return redirect_to @map, :status => :moved_permanently
-    end
+    redirect_to @map, status: :moved_permanently if @map && request.path != map_path(@map) && request.format == 'html'
   end
 
   def set_map
-
     # flexible query: Find slugged resource, if not, find a non-slugged resoure. All that to avoid an exception if a resource is unknown or not accessible (which would happen with friendly.find)
     @map = Map.by_user(current_user).find_by_slug(params[:id]) || Map.by_user(current_user).find_by_id(params[:id])
   end
 
   # Never trust parameters from the scary internet, only allow the white list through.
   def map_params
-    params.require(:map).permit(:title, :subtitle, :text, :published, :script, :group_id, :northeast_corner, :southwest_corner, :iconset_id, :basemap_url, :basemap_attribution)
+    params.require(:map).permit(:title, :subtitle, :text, :credits, :published, :script, :image, :group_id, :mapcenter_lat, :mapcenter_lon, :zoom, :northeast_corner, :southwest_corner, :iconset_id, :basemap_url, :basemap_attribution, :background_color, :popup_display_mode, :marker_display_mode, :show_annotations_on_map, :preview_url, :enable_privacy_features, :enable_map_to_go, :enable_historical_maps, :enable_time_slider)
   end
 end
