@@ -1,17 +1,26 @@
-function UserZoomLevel(map) {
-  map.on('zoomend',function(e){
-    var zl  = map.getZoom();
-    localStorage.setItem('UserZoomLevel',zl);
-  });
+/*jshint unparam: true, node: true */
+/*global window */
+/*global $ */
+
+function userZoomLevel(map) {
+    'use strict';
+    map.on('zoomend', function () {
+        var zl = map.getZoom();
+        localStorage.setItem('UserZoomLevel', zl);
+        var sw = screen.width;
+        var sw_em = parseInt(sw / parseFloat($('body').css('font-size')));
+        if ( sw_em > 39.9375 ) {
+            $('#show_zoomlevel').html("Level: " + zl);
+        }
+    });
 }
-function UserMapBounds(map) {
-  map.on('dragend',function(e){
-    var bounds = map.getBounds();
-    var northeast = bounds.getNorthEast();
-    var southwest = bounds.getSouthWest();
-    localStorage.setItem('UserMapBoundsNE',northeast['lat']+','+northeast['lng']);
-    localStorage.setItem('UserMapBoundsSW',southwest['lat']+','+southwest['lng']);
-
-  });
-
+function userMapBounds(map) {
+    'use strict';
+    map.on('dragend', function () {
+        var bounds = map.getBounds();
+        var northeast = bounds.getNorthEast();
+        var southwest = bounds.getSouthWest();
+        localStorage.setItem('UserMapBoundsNE', northeast.lat + ',' + northeast.lng);
+        localStorage.setItem('UserMapBoundsSW', southwest.lat + ',' + southwest.lng);
+    });
 }
