@@ -65,6 +65,12 @@ class Admin::UsersController < ApplicationController
     end
   end
 
+  protected
+
+  def update_resource(resource, account_update_params)
+    resource.update_without_password(account_update_params)
+  end
+
   private
 
   def admin_adresses
@@ -86,11 +92,5 @@ class Admin::UsersController < ApplicationController
     permitted_attributes = %i[email password group_id]
     permitted_attributes << :role if current_user.try(:admin?)
     params.require(:admin_user).permit(permitted_attributes)
-  end
-
-  protected
-
-  def update_resource(resource, params)
-    resource.update_without_password(params)
   end
 end
