@@ -10,14 +10,23 @@ module OrteBackend
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
+   
+    # Configuration for the application, engines, and railties goes here.
+    #
+    # These settings can be overridden in specific environments using the files
+    # in config/environments, which are processed later.
+    #
 
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # config.time_zone = "Central Time (US & Canada)"
+    # config.eager_load_paths << Rails.root.join("extras")
+
+    config.autoload_paths += %W(#{config.root}/lib)
+    
+    config.action_cable.mount_path = '/cable'
+
+    config.active_storage.replace_on_assign_to_many = false
 
     config.middleware.use Rack::Attack
-
-    config.active_record.sqlite3.represent_boolean_as_integer = true
 
     config.active_storage.content_types_to_serve_as_binary -= ['image/svg+xml']
   end
