@@ -1,11 +1,15 @@
-Rails.application.configure do
-  # Settings specified here will take precedence over those in config/application.rb.
+require "active_support/core_ext/integer/time"
 
   # The test environment is used exclusively to run your application's
   # test suite. You never need to work with it otherwise. Remember that
   # your test database is "scratch space" for the test suite and is wiped
   # and recreated between test runs. Don't rely on the data there!
-  config.cache_classes = true
+
+Rails.application.configure do
+  # Settings specified here will take precedence over those in config/application.rb.
+
+  config.cache_classes = false
+  config.action_view.cache_template_loading = true
 
   # Do not eager load code on boot. This avoids loading your whole application
   # just for the purpose of running a single test. If you are using a tool that
@@ -15,21 +19,25 @@ Rails.application.configure do
   # Configure public file server for tests with Cache-Control for performance.
   config.public_file_server.enabled = true
   config.public_file_server.headers = {
-    'Cache-Control' => "public, max-age=#{1.hour.seconds.to_i}"
+    'Cache-Control' => "public, max-age=#{1.hour.to_i}"
   }
 
   # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
-
-  # active storage
-  config.active_storage.service = :test
+  config.cache_store = :null_store
 
   # Raise exceptions instead of rendering exception templates.
   config.action_dispatch.show_exceptions = false
 
   # Disable request forgery protection in test environment.
   config.action_controller.allow_forgery_protection = false
+
+  # Store uploaded files on the local file system in a temporary directory.
+  config.active_storage.service = :test
+  # no, no:
+  # config.active_storage.variant_processor = :vips
+
   config.action_mailer.perform_caching = false
 
   # Tell Action Mailer not to deliver emails to the real world.
@@ -37,20 +45,30 @@ Rails.application.configure do
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
-  config.active_job.queue_adapter = :test
-
   # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr
 
+  # Raise exceptions for disallowed deprecations.
+  config.active_support.disallowed_deprecation = :raise
+
+  # Tell Active Support which deprecation messages to disallow.
+  config.active_support.disallowed_deprecation_warnings = []
+
+  # Raises error for missing translations.
+  # config.i18n.raise_on_missing_translations = true
+
+  # Annotate rendered view with file names.
+  # config.action_view.annotate_rendered_view_with_filenames = true
+
+  # rails5?
+  config.active_job.queue_adapter = :test
   config.action_mailer.raise_delivery_errors = false
-
   config.action_mailer.perform_deliveries = false
-
-  # Raises error for missing translations
-  # config.action_view.raise_on_missing_translations = true
-
   config.action_mailer.default_url_options = { :host => "test.domain.com" }
-
   routes.default_url_options[:host] = '127.0.0.1:3000'
+
+  # config.assets.compile = true
+  # config.action_controller.asset_host = "127.0.0.1:3000"
+  # config.assets.prefix = 'assets_test'
 
 end
