@@ -5,8 +5,8 @@ if Rails.env.production? || Rails.env.staging?
   Rails.application.config.middleware.use ExceptionNotification::Rack,
     email: {
       email_prefix:           "[ORTE Backend - #{::Rails.env}] ",
-      sender_address:         %{ technik@3plusx.de  },
-      exception_recipients:   %w{ technik@3plusx.de   }
+      sender_address:         %{ dev@a-thousand-channels.xyz  },
+      exception_recipients:   %w{ dev@a-thousand-channels.xyz }
   }
 
 end
