@@ -2,6 +2,6 @@
 #  DEPLOY_REVISION =  File.read(File.join(Rails.root, 'REVISION'))[0..7]
 #  DEPLOY_DATE = File.new(File.join(Rails.root, 'REVISION')).ctime
 # else
-  DEPLOY_REVISION = "local"
+  DEPLOY_REVISION = ""
   DEPLOY_DATE = DateTime.now
 # end
