@@ -5,7 +5,7 @@ require 'rails_helper'
 RSpec.describe Public::MapsController, type: :controller do
   render_views
 
-  describe "functionalities with logged in user with role 'admin'" do
+  describe 'functionalities with anonymous access' do
     before do
       @group = FactoryBot.create(:group)
       user = FactoryBot.create(:admin_user, group_id: @group.id)
@@ -26,16 +26,10 @@ RSpec.describe Public::MapsController, type: :controller do
     let(:valid_session) { {} }
 
     describe 'GET #index' do
-      it 'returns a success response for published maps' do
+      it 'returns a 403' do
         map = Map.create! valid_attributes
         get :index, params: { format: 'json' }, session: valid_session
-        expect(response).to have_http_status(200)
-      end
-
-      it 'returns a success response for unpublished resources' do
-        map = Map.create! invalid_attributes
-        get :index, params: { format: 'json' }, session: valid_session
-        expect(response).to have_http_status(200)
+        expect(response).to have_http_status(403)
       end
     end
 
@@ -51,7 +45,7 @@ RSpec.describe Public::MapsController, type: :controller do
         get :show, params: { id: map.friendly_id, format: 'json' }, session: valid_session
         expect(response).to have_http_status(200)
         expect(response.status).to eq 200
-        expect(response.content_type).to eq('application/json')
+        expect(response.content_type).to eq('application/json; charset=utf-8')
       end
 
       it 'returns json with a valid scheme' do
@@ -59,13 +53,49 @@ RSpec.describe Public::MapsController, type: :controller do
         layer = FactoryBot.create(:layer, map_id: map.id, published: true)
         place = FactoryBot.create(:place, layer_id: layer.id, published: true)
         get :show, params: { id: map.friendly_id, format: 'json' }, session: valid_session
-        # puts response.body
         expect(response).to match_response_schema('map', 'json')
       end
 
-      xit 'returns an 403 + error response for unpublished resources' do
+      it 'returns an 403 + error response for unpublished resources' do
         map = Map.create! invalid_attributes
         get :show, params: { id: map.friendly_id, format: 'json' }, session: valid_session
+        expect(response).to have_http_status(403)
+        expect(JSON.parse(response.body)['error']).to match(/Map not accessible/)
+      end
+    end
+
+    describe 'GET #allplaces' do
+      it 'returns a success response for a published map' do
+        map = Map.create! valid_attributes
+        layer = FactoryBot.create(:layer, map_id: map.id, published: true)
+        place = FactoryBot.create(:place, layer_id: layer.id, published: true)
+        get :allplaces, params: { id: map.friendly_id, format: 'json' }, session: valid_session
+        expect(response).to have_http_status(200)
+      end
+
+      it 'returns json for a published map' do
+        map = Map.create! valid_attributes
+        layer = FactoryBot.create(:layer, map_id: map.id, published: true)
+        place = FactoryBot.create(:place, layer_id: layer.id, published: true)
+        get :allplaces, params: { id: map.friendly_id, format: 'json' }, session: valid_session
+        expect(response).to have_http_status(200)
+        expect(response.status).to eq 200
+        expect(response.content_type).to eq('application/json; charset=utf-8')
+      end
+
+      it 'returns json with a valid scheme' do
+        map = Map.create! valid_attributes
+        layer = FactoryBot.create(:layer, map_id: map.id, published: true)
+        place = FactoryBot.create(:place, layer_id: layer.id, published: true)
+        get :allplaces, params: { id: map.friendly_id, format: 'json' }, session: valid_session
+        expect(response).to match_response_schema('map_allplaces', 'json')
+      end
+
+      it 'returns an 403 + error response for unpublished resources' do
+        map = Map.create! invalid_attributes
+        layer = FactoryBot.create(:layer, map_id: map.id, published: true)
+        place = FactoryBot.create(:place, layer_id: layer.id, published: true)
+        get :allplaces, params: { id: map.friendly_id, format: 'json' }, session: valid_session
         expect(response).to have_http_status(403)
         expect(JSON.parse(response.body)['error']).to match(/Map not accessible/)
       end
