@@ -5,9 +5,20 @@ FactoryBot.define do
     title { 'MyString' }
     subtitle { 'MyString' }
     text { 'MyString' }
+    teaser { 'MyTeaser' }
+    credits { 'MyString' }
+    style { 'MyCSS' }
     published { false }
-    basemap_url { 'MyBasemapUrl' }
+    basemap_url { 'MyMapBasemapUrl' }
     basemap_attribution { 'Basemap made by' }
+    color { '#0000cc' }
+    background_color { '#151515' }
+    mapcenter_lat { '0.1' }
+    mapcenter_lon { '10' }
+    zoom { 12 }
+    popup_display_mode { 'click' }
+    tooltip_display_mode { 'false' }
+    places_sort_order { 'startdate' }
     group
     trait :invalid do
       title { nil }
