@@ -17,8 +17,12 @@ RSpec.describe PlacesHelper, type: :helper do
 
   describe 'image_link' do
     it 'it returns an polymorphic image link' do
-      i = Image.new
-      i.attach(io: File.open(Rails.root.join('spec', 'support', 'files', 'test.jpg')), filename: 'attachment.jpg', content_type: 'image/jpeg')
+      p = create(:place)
+      i = build(:image, place: p)
+      uploaded = Rack::Test::UploadedFile.new(Rails.root.join('spec', 'support', 'files', 'test.jpg'), 'image/jpeg')
+      i.attach(uploaded)
+      i.save!
+      i.reload
       expect(helper.image_link(i)).to eq("http://test.host#{polymorphic_path(i.file.variant(resize: '800x800').processed)}")
     end
   end
@@ -26,7 +30,7 @@ RSpec.describe PlacesHelper, type: :helper do
   describe 'icon_link' do
     it 'it returns an polymorphic icon link' do
       i = Icon.new
-      i.file.attach(io: File.open(Rails.root.join('spec', 'support', 'files', 'test.jpg')), filename: 'attachment.jpg', content_type: 'image/jpeg')
+      i.file.attach(io: File.open(Rails.root.join('spec', 'support', 'files', 'test.jpg')), filename: 'attachment.jpg', content_type: 'image/jepg')
       expect(helper.icon_link(i.file)).to eq("<img src=\"http://test.host#{rails_blob_path(i.file)}\">")
     end
   end
@@ -46,8 +50,9 @@ RSpec.describe PlacesHelper, type: :helper do
   describe 'audio_link' do
     it 'it returns an polymorphic audio link' do
       p = Place.new
-      p.audio.attach(io: File.open(Rails.root.join('spec', 'support', 'files', 'test.mp3')), filename: 'attachment.mp3', content_type: 'audio/mpeg')
-      expect(helper.audio_link(p.audio)).to eq("<audio autoplay=\"autoplay\" controls=\"controls\" src=\"http://test.host#{rails_blob_path(p.audio)}\"></audio>")
+      uploaded = Rack::Test::UploadedFile.new(Rails.root.join('spec', 'support', 'files', 'test.mp3'), 'audio/mpeg')
+      p.audio.attach(uploaded)
+      expect(helper.audio_link(p.audio)).to eq("<audio controls=\"controls\" src=\"http://test.host#{rails_blob_path(p.audio)}\"></audio>")
     end
   end
 end
